refactor: 优化组件目录结构
