+++ components.d.ts
@@ -8,13 +8,13 @@ export {}
 /* prettier-ignore */
 declare module 'vue' {
   export interface GlobalComponents {
-    Badge: typeof import('./src/components/Badge.vue')['default']
+    Badge: typeof import('./src/components/common/Badge.vue')['default']
     BannerTitle: typeof import('./src/components/BannerTitle.vue')['default']
     ContactCard: typeof import('./src/components/ContactCard.vue')['default']
     DashboardCard: typeof import('./src/components/DashboardCard.vue')['default']
     Explanations: typeof import('./src/components/Explanations.vue')['default']
     FloatingImages: typeof import('./src/components/FloatingImages.vue')['default']
-    Footer: typeof import('./src/components/Footer.vue')['default']
+    Footer: typeof import('./src/components/common/Footer.vue')['default']
     IIcRoundWechat: typeof import('~icons/ic/round-wechat')['default']
     IMaterialSymbolsBook2Rounded: typeof import('~icons/material-symbols/book2-rounded')['default']
     IMaterialSymbolsCircle: typeof import('~icons/material-symbols/circle')['default']
@@ -36,21 +36,21 @@ declare module 'vue' {
     ISimpleIconsXiaohongshu: typeof import('~icons/simple-icons/xiaohongshu')['default']
     ISimpleIconsZhihu: typeof import('~icons/simple-icons/zhihu')['default']
     LetterPaper: typeof import('./src/components/LetterPaper.vue')['default']
-    Link: typeof import('./src/components/Link.vue')['default']
+    Link: typeof import('./src/components/common/Link.vue')['default']
     NavBar: typeof import('./src/components/NavBar.vue')['default']
     NavPanel: typeof import('./src/components/NavPanel.vue')['default']
     PageContent: typeof import('./src/components/PageContent.vue')['default']
     Pagination: typeof import('./src/components/Pagination.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
-    RubyText: typeof import('./src/components/RubyText.vue')['default']
+    RubyText: typeof import('./src/components/common/RubyText.vue')['default']
     SearchBar: typeof import('./src/components/SearchBar.vue')['default']
     SearchSkeleton: typeof import('./src/components/SearchSkeleton.vue')['default']
-    SeeSymbol: typeof import('./src/components/SeeSymbol.vue')['default']
+    SeeSymbol: typeof import('./src/components/common/SeeSymbol.vue')['default']
     SideBar: typeof import('./src/components/SideBar.vue')['default']
-    SpeechBubble: typeof import('./src/components/SpeechBubble.vue')['default']
+    SpeechBubble: typeof import('./src/components/common/SpeechBubble.vue')['default']
     SpeechBubbles: typeof import('./src/components/SpeechBubbles.vue')['default']
-    Subtitle: typeof import('./src/components/Subtitle.vue')['default']
+    Subtitle: typeof import('./src/components/common/Subtitle.vue')['default']
     Timeline: typeof import('./src/components/Timeline.vue')['default']
     WordCikLingCard: typeof import('./src/components/WordCikLingCard.vue')['default']
     WordCikLingEntry: typeof import('./src/components/WordCikLingEntry.vue')['default']
