+++ src/components/common/Link.vue
