+++ src/components/common/SeeSymbol.vue
