+++ src/components/common/SpeechBubble.vue
